Load data for ablanor

Functions for loading data needed by ablanor from a database or local files. When local files are used the environmental variable filbane_ablanor_test must be set to the path of the file where data are found

getDataDump(registryName, tableName, fromDate, toDate, ...)

getNameReshId(registryName, asNamedList = FALSE)

getHospitalName(registryName, reshId)

getRand12(registryName, singleRow, ...)

getProsPatient(registryName, singleRow, ...)

Arguments

registryName
Character string defining the registry name.

tableName
Character string with name of database table

fromDate
Character string of format YYYY-MM-DD with start date

toDate
Character string of format YYYY-MM-DD with end date

...
Optional arguments to be passed to the function.

asNamedList
Logical wether to return a list of named values or not. Default is FALSE in which case a data frame containing name and id is returned.

reshId
Integer organization id.

singleRow
Logical if only one row from the table is to be provided. Default value is FALSE.

Value

Data frame or (when multiple data sets are returned) a list of data frames containing registry data. In case of getNameReshId() data may also be returned as a named list of values (see Details).

Details

getNameReshId() returns a mapping of organization name and id in the form of columns named name and id. Optionally this function can also return a list of named values (ids), e.g. for use in shiny selection lists.