Ablanor SQL

Get tables from database usig SQL. Notice, only tables when a procedure-date exists!

getBasereg(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getPros(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getMce(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getRand12(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupBasis(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupOneYr(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFollowupFiveYr(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getGkv(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getProms(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getHendelse(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getFriendlycentre(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getMcepatientdata(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate,
  toDate,
  ...
)

getPatientlist(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregPros(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsFollowup1(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsFollowup0(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getBaseregProsFollowup5(
  registryName,
  singleRow,
  reshId = NULL,
  userRole,
  fromDate = NULL,
  toDate = NULL,
  ...
)

getLatestEntry(registryName)

getNameReshId(
  registryName,
  asNamedList = FALSE,
  shortNames = FALSE,
  newNames = FALSE
)

getHospitalName(registryName, reshId, shortName = FALSE, newNames = FALSE)

Arguments

registryName

"ablanor"

singleRow

bools. TRUE only one row, for variable names. FALSE all rows.

reshId

Integer organization id. From login settings.

userRole

String dummy/placeholder role. "LC" has access only to local data (defined by reshId), "SC" has access to national data.

fromDate

NULL default is 01-01-1900. If datadump or pivot table, start date of calendar is used.

toDate

NULL default is newest registration in Abalnor. If datadump or pivot table, end date of calendar is used.

...

Optional arguments to be passed to the function.

asNamedList

Logical whether to return a list of named values or not. Default is FALSE in which case a data frame containing name and id is returned.

newNames

boolean. TRUE uses "sykehusnavn" as defined in 'legg_til_sykehusnavn()'. Default value is FALSE, uses "sykehusnavn" from table Friendlycentre.

shortName

boolean. Default value FALSE and "friendlyname" is returned. If TRUE shortname is returned.

Value

Data frame or (when multiple data sets are returned) a list of data frames containing registry data. In case of getNameReshId() data may also be returned as a named list of values (see Details).

Details

getNameReshId() returns a mapping of organization name and id in the form of columns named name and id. Optionally this function can also return a list of named values (ids), e.g. for use in shiny selection lists.