Utvid og organiser Funksjon som utvider tabellen og legger til alle manglende kombinasjoner av variablene "maned" og "legend". Deretter brukes tidyr::pivot_wider for å få en rad per måned og en kolonne per nivå av faktoren "legend"

Utvid og organiser Funksjon som utvider tabellen og legger til alle manglende kombinasjoner av variablene "maned" og "legend". Deretter brukes tidyr::pivot_wider for å få en rad per måned og en kolonne per nivå av faktoren "legend"

leggTilTomme_PivotWide(data, remove_NAcols = TRUE)

Arguments

data

datasett med variablene maaned,legend og value

remove_NAcols

Boolsk variabel. Remove_NA =TRUE dersom nivået "NA" skal fjernes. FALSE, dersom vi ønsker beholde nivået "NA" og ha en kolonne for dette nivået

Value

data.frame med en rad per nivå av maaned og en kolonne per nivå av "legend"

Examples

x <- data.frame(maaned = rep(c("jan", "feb", "mars", "april"), 2),
  legend = factor(x = c(rep("en",4) ,rep("to",4)),
  levels = c("en", "to", "tre")),
  value = paste0("1/", 1:8))
leggTilTomme_PivotWide(x, remove_NA = FALSE)
#> # A tibble: 4 × 4
#>   maaned en    to    tre   
#>   <chr>  <chr> <chr> <chr> 
#> 1 april  1/4   1/8   " -- "
#> 2 feb    1/2   1/6   " -- "
#> 3 jan    1/1   1/5   " -- "
#> 4 mars   1/3   1/7   " -- "