Common report processor for AblaNor

Makes reports for AblaNor typically used for auto reports such as subscriptions, dispatchments and bulletins. As such, please be warned that any changes to this function might render existing auto reports nonfunctional as they are based on static calls based on any previous version of this function. Changes should therefore be thoroughly tested against existing auto reports. Altering the names of the arguments will likely be a breaking change. Adding new arguments should be safe as long as they are provided a default value.

reportProcessor(
  report,
  outputType = "pdf",
  title = "",
  author = "unknown author",
  orgName = "unknown organization",
  orgId = 999999,
  registryName = "ablanor",
  userFullName = "unknown person name",
  userRole = "unknown role",
  userOperator = "unknown operator"
)

Arguments

report

Character string identifying the report to be processed by this function.

outputType

Character string with output format. Must be one of c("html", "pdf"). Defaults to "pdf".

title

Character string giving the report title. Empty string by default.

author

Character string providing the report author. Default value is "unknown author".

orgName

Character string with the name of the organization/hospital. Default is "unknown organization".

orgId

Integer (?) with the id of the organization/hospital. Default is 999999.

registryName

Character string with registry name. Default is "ablanor".

userFullName

Character string giving the person name, normally the user requesting the report. Default is "unknown person name".

userRole

Character string giving a user role, normally the one of the user requesting the report. Default is "unknown role".

userOperator

Character string with some name of an operator, whatever that is... Default is "unknown operator".

Value

A character string with a path to where the produced file is located.

Examples

## Make the start page for ablanor
if (FALSE) reportFilePath <- reportProcessor(report = "veiledning",
                                  title = "Example report")
 # \dontrun{}